add react-query

+++ src/utils/RestApiUtil.ts
@@ -1,9 +1,26 @@
 import axios, { AxiosInstance, InternalAxiosRequestConfig } from 'axios';
 
-export interface CommonResponse<T = any> {
+export interface RestApiRequest {
+  url: string;
+  method: Method;
+  service: string; //Service;
+  queryParams?: URLSearchParams;
+  bodyParams?: object;
+  headers?: object;
+  responsType?: XMLHttpRequestResponseType;
+}
+
+export interface RestApiResponse<T = any> {
   successOrNot: string;
   statusCode: string;
   data?: T;
+}
+export enum Method {
+  GET = 'GET',
+  POST = 'POST',
+  PUT = 'PUT',
+  DELETE = 'DELETE',
+  PATCH = 'PATCH',
 }
 export enum Service {
   AQUA_BE = 'aqua-be',
@@ -14,11 +31,11 @@ export enum ServicePort {
   AQUA_ADMIN_BE = 7071,
 }
 
-const getInstance = (serviceName: string, params?: any): AxiosInstance => {
+const getInstance = (request: RestApiRequest): AxiosInstance => {
   axios.defaults.headers.post['Content-Type'] = 'application/json';
 
   let baseURL = '';
-  switch (serviceName) {
+  switch (request.service) {
     case Service.AQUA_BE:
       if (process.env.REACT_APP_NODE_ENV === 'local') {
         baseURL = `${process.env.REACT_APP_API_BASE_URL}:${ServicePort.AQUA_BE}`;
@@ -28,14 +45,41 @@ const getInstance = (serviceName: string, params?: any): AxiosInstance => {
       baseURL = `${process.env.REACT_APP_API_BASE_URL}:${ServicePort.AQUA_ADMIN_BE}`;
       break;
     default:
+      baseURL = `${request.service}${request.url}`;
       break;
   }
 
   // axios 인스턴스 생성
-  const instance = axios.create({
-    baseURL: baseURL,
-    timeout: 1000,
-  });
+  // 메소드별 인스턴스 설정, request 설정이 없으면 default 값
+  let config = {};
+  switch (request.method) {
+    case Method.GET:
+    case Method.DELETE:
+      config = {
+        baseURL: baseURL,
+        timeout: 3000,
+        headers: request?.headers || {
+          Accept: 'application/json',
+        },
+        params: request?.queryParams || {},
+        responseType: request?.responsType || 'json',
+      };
+      break;
+    case Method.POST:
+    case Method.PATCH:
+    case Method.PUT:
+      config = {
+        baseURL: baseURL,
+        timeout: 1000,
+        headers: request?.headers || {
+          Accept: 'application/json',
+          'Content-Type': 'application/json',
+        },
+        responseType: request?.responsType || 'json',
+      };
+      break;
+  }
+  const instance = axios.create(config);
 
   instance.interceptors.request.use(
     (config: InternalAxiosRequestConfig) => {
@@ -54,17 +98,28 @@ const getInstance = (serviceName: string, params?: any): AxiosInstance => {
 
   instance.interceptors.response.use(
     async (response: any): Promise<any> => {
-      const commonResponse: CommonResponse = response.data as CommonResponse;
+      //응답값이 file이라면, 나중에 util 분리
+      if (response.data instanceof Blob) {
+        const CommonBlobResponse: RestApiResponse<Blob> = {
+          successOrNot: 'Y',
+          statusCode: 'Success',
+          data: new Blob([response?.data]),
+        };
+
+        return CommonBlobResponse;
+      }
+
+      const commonResponse: RestApiResponse = response.data as RestApiResponse;
 
       if (commonResponse.statusCode && commonResponse.statusCode === 'SESSION_EXPIRE') {
         sessionStorage.clear();
-        window.location.assign('/login');
+        window.location.assign('/login'); //컴포넌트가 아니면 router 라이브러리 사용이 불가하기 때문
       }
       return commonResponse;
     },
 
     async (error: any): Promise<any> => {
-      const unknownError: CommonResponse = {
+      const unknownError: RestApiResponse = {
         successOrNot: 'N',
         statusCode: 'UNKNOWN_ERROR',
         data: {},
@@ -80,26 +135,11 @@ const getInstance = (serviceName: string, params?: any): AxiosInstance => {
   return instance;
 };
 
-export enum Method {
-  GET = 'GET',
-  POST = 'POST',
-  PUT = 'PUT',
-  DELETE = 'DELETE',
-  PATCH = 'PATCH',
-}
-
 export interface QueryParams {
   [key: string]: string | number | boolean;
 }
 
-export interface ApiRequest {
-  service: string;
-  url: string;
-  method: Method;
-  queryParams?: QueryParams;
-  bodyParams?: object;
-}
-
+//미사용
 const getQueryStringFormat = (queryParams?: QueryParams): string => {
   if (!queryParams) return '';
   const keys = Object.keys(queryParams);
@@ -110,10 +150,10 @@ const getQueryStringFormat = (queryParams?: QueryParams): string => {
   return queryString ? `?${queryString}` : '';
 };
 
-export const callApi = async (apiRequest: ApiRequest): Promise<CommonResponse> => {
-  const url: string = apiRequest.url + getQueryStringFormat(apiRequest?.queryParams);
+export const callApi = async (apiRequest: RestApiRequest): Promise<RestApiResponse> => {
+  //const url: string = apiRequest.url + getQueryStringFormat(apiRequest?.queryParams);
 
-  let response: CommonResponse = {
+  let response: RestApiResponse = {
     successOrNot: 'N',
     statusCode: 'UNKNOWN_ERROR',
     data: {},
@@ -121,19 +161,19 @@ export const callApi = async (apiRequest: ApiRequest): Promise<CommonResponse> =
 
   switch (apiRequest.method) {
     case Method.GET:
-      response = await getInstance(apiRequest.service).get(url);
+      response = await getInstance(apiRequest).get(apiRequest.url);
       break;
     case Method.POST:
-      response = await getInstance(apiRequest.service).post(url, apiRequest?.bodyParams);
+      response = await getInstance(apiRequest).post(apiRequest.url, apiRequest?.bodyParams || {});
       break;
     case Method.PUT:
-      response = await getInstance(apiRequest.service).put(url, apiRequest?.bodyParams);
+      response = await getInstance(apiRequest).put(apiRequest.url, apiRequest?.bodyParams || {});
       break;
     case Method.DELETE:
-      response = await getInstance(apiRequest.service).delete(url);
+      response = await getInstance(apiRequest).delete(apiRequest.url);
       break;
     case Method.PATCH:
-      response = await getInstance(apiRequest.service).patch(url, apiRequest?.bodyParams);
+      response = await getInstance(apiRequest).patch(apiRequest.url, apiRequest?.bodyParams || {});
       break;
     default:
       break;
